+++ CHANGELOG.md
@@ -3,6 +3,7 @@ yii2-dynamicform change Log
 
 dev-master
 ----------
+- enh: Added support for "kartik-v/yii2-widget-depdrop"
 - enh: Added support for "kartik-v/yii2-widget-select2"
 - bug: Fixed html name attribute for "kartik-v/yii2-widget-colorinput"
 - enh: Added support for "kartik-v/yii2-widget-timepicker"